removed cache and refactored
